arch(lsp): split server.go into focused sibling files
server.go was 1007 lines. Split the session management and utility
helpers into two new sibling files within the lsp package:

- server_session.go: rebuildSession, currentSession, snapshotConfig,
  reloadConfig, resolveConfig, fetchClientSettings, registerPendingResponse,
  unregisterPendingResponse, deliverResponse
- server_util.go: workspaceRelative, dirFSForPath, frontMatterEnabled,
  isFixable, uriToPath, uriToPathOnOS, hasDriveLetterPrefix, pickRoot

server.go trimmed to 555 lines (core struct, types, New, Run,
recoverPanic, logPanic, and the dispatch functions). Pure reorganisation
— zero logic changes.

Closes plan/2606211909_arch-fix-lsp-server-split.md.
